requirements.txt
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cloudinary==1.29.0
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
fcm-django==1.0.8
firebase-admin==5.2.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.3
packaging==21.3
Pillow==9.0.1
proto-plus==1.20.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
